Refactor plot_model and make it look nicer by default
Adds a band_xlim kwarg that can be used to configure the range of the likelihood bands in the left two plots.
The subplots are now created with a fine tuned layout that is square and has nice looking margins inbetween the plots.
Log-scale is automatically turned off when calibration data is <= 0.
Closes #47